[BUGFIX] Set opt-in for override checkbox in TCA input elements
When useOrOverridePlaceholder mode is used for TCA input type,
use opt-in for the checkbox when placeholder is empty.

In some cases, the placeholder can be empty. Especially when the
placeholder uses data from another field. When information is added
to the corresponding field, the input field does not take this in account
because the override checkbox is set by default.

This can lead to problems for title and alternative fields in the
sys_file_metadata table. When adding a file reference and setting the
title or alternative field afterwards in the original file, the reference
is still using the empty override from file reference.

Resolves: #84563
Releases: master, 8.7
Change-Id: I84a37d55a48464b6249fc8bcee405a7cd909016a
Reviewed-on: https://review.typo3.org/56989
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>