[FEATURE] Store icons in localStorage
The icons that get requested by the Icon API on JavaScript side are not
stored in the client's localStorage.
To have a proper invalidation, a hash of the IconRegistry is built and
stored in the localStorage, too. If the hash changes, all icons in the
localStorage get flushed.

To achieve this, the Storage/Client module is extended to allow removing
values by a given prefix.

Resolves: #84780
Releases: master
Change-Id: Ic2137b05530201a8a94a7ea6c28ae1a012206221
Reviewed-on: https://review.typo3.org/56721
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Kay Strobach <typo3@kay-strobach.de>
Tested-by: Kay Strobach <typo3@kay-strobach.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>