[BUGFIX] Language selector not shown in default page without translation
When editing a page, the language selector in the upper left corner is
not visible when the page is in the default language and does not have a
translation yet. When the page has at least one translation, the language
selector is visible, which is inconsistent behaviour.

This patch adds a check for the table. When the languages for the page
table are fetched, all languages are fetched.

Resolves: #84834
Releases: master
Change-Id: Ib482fecef6cbddcbf945e0e4d944e181c27a115c
Reviewed-on: https://review.typo3.org/56774
Reviewed-by: Sebastian Hofer <sebastian.hofer@marit.ag>
Tested-by: Sebastian Hofer <sebastian.hofer@marit.ag>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Tested-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Christian Hellmund <developement@hellmund.eu>
Reviewed-by: Florian Mast <flo.mast@web.de>
Tested-by: Florian Mast <flo.mast@web.de>
Reviewed-by: Ruslan Steiger <steiger.ruslan@gmail.com>
Tested-by: Ruslan Steiger <steiger.ruslan@gmail.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>