[BUGFIX] Only query default language PIDs in suggest wizard
Due to the removal of "pages_language_overlay", the page IDs for
FormEngine's suggest wizard contains page IDs of all languages, however
it is defined that the "pid" field is only limited to the default language.

Adding a check for `sys_language_uid=0` ensures that the pid IN
query is much smaller than before.

Resolves: #85308
Releases: master
Change-Id: I21e2c706dc3d73986e027cb9e014e3f69980d3ba
Reviewed-on: https://review.typo3.org/57265
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>