[BUGFIX] Move page correctly when droping page before another one
Fixes an issue where drag and dropping page before a another one didn't
detect the previous page correctly. DataHandler always expect
us to put a node AFTER. So when we want to place a page before, we need
to find previous page and put the new on after it.

Now it correctly compares depth of the node mouse is over with a previous
node. Before it took a dragged node depth for a comparison, which was wrong.

Releases: master
Resolves: #84008
Change-Id: Ic43be1147a5bb0fc1e15943a5d9e235d10727a07
Reviewed-on: https://review.typo3.org/57809
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Kummer <typo3@enobe.de>
Tested-by: Joerg Kummer <typo3@enobe.de>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>