[TASK] Centralize sys_domain resolving
Core provides various places to either retrieve the pageId
of a sys_domain record, or to find out the root page ID
given a request (previously domain + path).

However, due to the inclusion of symfony/routing, the
detection can now be handled by this new component.

For this purpose, a new class LegacyDomainResolver is
introduced (internal) which utilizes symfony/routing
under the hood and does proper caching.

This way, the following methods can be deprecated:
- TypoScriptFrontendController->domainNameMatchesCurrentRequest()
- TypoScriptFrontendController->getDomainDataForPid()
- BackendUtility::getDomainStartPage()
- BackendUtility::firstDomainRecord()

This is a precursor for fetching all sys_domain data,
to minimize the efforts later-on to wrap the LegacyDomainResolver
into a proper Router, and also to build pseudo-sites
on top.

Resolves: #85892
Releases: master
Change-Id: I831d33ac06f090cbe4d2a16e592549739489f99a
Reviewed-on: https://review.typo3.org/57949
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>