[TASK] Cleanup unused PageUriBuilder options
Due to the refactoring of the typolink method, the PageUriBuilder
is now only used for links to pages with a site configuration, so it
can be cleaned up.

This a first step in getting the PageUriBuilder API in shape.

As this class is new, and the options and the class is not public
API yet, this removal does not need any docs.

Resolves: #86052
Releases: master
Change-Id: I56205f4e3fec0e68bd317a45ae1d860f8d84a799
Reviewed-on: https://review.typo3.org/58095
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>