[BUGFIX] Expression Language should not crash Backend Modules
When doing "[page["uid"] >= 284]" in a TypoScript condition, everything works
as expected in Frontend, but in BE it gets messy, as "page" HAS to be an array
to allow to validate this, but page does not always exist.

This comes from a wrong spec, claiming that conditionMatcher->getPage() always
returns an array, which is wrong. So let's make the code return an array all the time.

Resolves: #86062
Releases: master
Change-Id: Ia781c84ffa335fce5ba2d67bbe2997ceab5281b4
Reviewed-on: https://review.typo3.org/58102
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>