[BUGFIX] Properly create slug if adding new pages via FormEngine
Adding a page relative to another page (vanillaUid is negative then),
currently creates a slug as if the page in inside the related page.
This is because the vanillaUid is given to the JS and later fed to
the ajax slug creator.
Using 'effectivePid' instead which is the parent page uid if a
new page is added, or the uid of the page recod itself if the
page is edited, solves the issue.

Resolves: #86074
Releases: master
Change-Id: Ia3f89f615699d4e7be6bc1762bc95e17cc1939dc
Reviewed-on: https://review.typo3.org/58117
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>