[TASK] Deprecate various methods inside BE_USER
The property "checkWorkspaceCurrent_cache" is now protected.

Second method argument of "modAccess()" is deprecated.

"isPSet()" is not in use anymore, and will be removed in TYPO3 v10.0.

Some other methods, which are workspace-related, are marked
as internal in order to properly move them into EXT:workspaces
in the next version.

Resolves: #86441
Releases: master
Change-Id: Ie776e71cdc7485937b2a38e7cfc93c13aba925ab
Reviewed-on: https://review.typo3.org/58458
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>