[TASK] Add proper exception handling to RouterInterface logic
Two new Exceptions are now thrown when routing does not
work, one being thrown when a URL is generated but could
not be generated, and one when a URL could not be resolved.

This is much cleaner than the distinction of a nullable return type,
so the new interface is adapted as well.

As a drive-by fix, the Backend routing exception now inherits from
this new exception.

Resolves: #86500
Releases: master
Change-Id: Ifaf7b61422dfd49df18399c3bcbdf735bc522cba
Reviewed-on: https://review.typo3.org/58519
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>