[BUGFIX] Omit `overrideVals` in NewRecordController
In the old times of pages_language_overlay, the "New Record" wizard
automatically re-used the doktype of the parent table which was `pages`.
Since pages_language_overlay vanished, the `overrideVals` in the links in
the wizard use the `pages` table.

This results in invalid value for `doktype` if a page in pid=0 is
created, as pid=0 has no doktype.

For this reason, the `overrideVals` parameter is removed.

Resolves: #86547
Related: #82445
Releases: master
Change-Id: I6efa699e5ef09eed6321ae68b5838682214376de
Reviewed-on: https://review.typo3.org/58561
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>