[BUGFIX] Display FlashMessage in View Module
When clicking on the root page while in view module
a FlashMessage is now displayed with proper information.
The FlashMessage is also shown in the case of clicking the
root page and a Page, marked as siteroot, exists.
Furthermore the FlashMessage is now also shown if the
selected page has no valid doktype e.g. sysfolder
or spacer.

Resolves: #86563
Releases: master
Change-Id: Ia8c6cddfe38602c94b41d5da81987d061778096f
Reviewed-on: https://review.typo3.org/58573
Reviewed-by: Rudy Gnodde <rgn@windinternet.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Jasmina Ließmann <code@frauliessmann.de>
Reviewed-by: Joerg Kummer <typo3@enobe.de>
Tested-by: Joerg Kummer <typo3@enobe.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>