[TASK] Improve click area of new content element wizard item
The clickable area of a new content element wizard item
is improved now.

The <a>-tag was added around the Bootstrap 3 media object
childrens which makes it easier now to add a new content
element by clicking anywhere on the item.

The text shows also the hover-style when an editor hovers
the icon of a NCEW item. The user experience for editors
working on touch devices is also improved by this.

Resolves: #86590
Releases: master
Change-Id: Ib6dc65a0d926945e6e9dd8a9fe7b57e5ee27d8f9
Reviewed-on: https://review.typo3.org/58594
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>