[BUGFIX] Respect saveDocNew TSConfig option
For translated tt_content elements the saveDocNew button
was solely dependent on the free vs. connected mode setup.
However, if the button has been hidden via TSConfig
completely it should not magically reappear when editing a
translation in free mode.

(This patch only applies to TYPO3 8 as the buttons
changed for version 9).

Change-Id: I0f74e87acbac127b8bbc09f4fee39c633a3a49f6
Resolves: #85774
Releases: 8.7
Reviewed-on: https://review.typo3.org/58603
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Joerg Kummer <typo3@enobe.de>
Tested-by: Joerg Kummer <typo3@enobe.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>