[BUGFIX] Pass status code as int to RedirectResponse
The filelist module shows an error when creating an new textfile
with a file extension that is not allowed via [SYS][textfile_ext],
e.g. test.md and sends an according RedirectResponse.

The RedirectResponse constructor call has been fixed to be called with
an int instead of a string as status code.

Resolves: #86765
Releases: master
Change-Id: I63ea741223db21b8e4f8fbda1cc958b608ff981e
Reviewed-on: https://review.typo3.org/58731
Reviewed-by: Joerg Kummer <typo3@enobe.de>
Tested-by: Joerg Kummer <typo3@enobe.de>
Reviewed-by: Felix Althaus <felix.althaus@undkonsorten.com>
Tested-by: Felix Althaus <felix.althaus@undkonsorten.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>