[SECURITY] Avoid disclosing loaded extensions
Inline JavaScript settings for RequireJS and ajaxUrls disclose the
existence of specific extensions in a TYPO3 installation.

In case no backend user is logged in RequireJS settings are fetched
using an according endpoint, ajaxUrls (for backend AJAX routes) are
limited to those that are accessible without having a user session.

Resolves: #83855
Releases: master, 9.5, 8.7
Security-Commit: af76c928bbe6fe05611db0839da879fce132daff
Security-Bulletin: TYPO3-CORE-SA-2019-001
Change-Id: I90dddd2fd3fd81834cd40c8638fa487fa106b07c
Reviewed-on: https://review.typo3.org/59520
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>