[BUGFIX] Properly show unchanged workspace records in workspace module
When a user is editing a complex data structure with IRRE and
if he just changes one entry of a structure then
all the items will be shown as changed in the workspace module.
Most users will wonder why an element is shown as changed even
if the user didn't touch this particular entry.

This change will hide the "show differences" button in the module
for entries where no data was changed except for internal properties.

Those items will also have a different color in the list. Now the user
can easily identify the records where something was really
changed and open the diff to see what has changed.

Resolves: #91157
Releases: master
Change-Id: Ib5ca5bb50360091813855bd8aef0b9e627e1e826
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64295
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>