[TASK] Switch functional tests to site handling
All functional tests that run a frontend subrequest
should be tested against Site Handling, as the
PseudoSite Handling is about to removed.

This patch is actually a precursor to make sure
that all tests work with site handling now.

The AbstractDataHandlerActionTestCase adds
e site configuration on each setUp(). This
way, all DataHandler-related tests are running
with a Site configuration of three languages.

In addition, when "localizing" something,
a translated page needs to exist, so most
DataHandler related changes add this functionality.

For some DataHandler tests, the SiteConfiguration
is adapted to define the previous behavior
built in TypoScript.

One addition in PageResolver is made:
If an GET / POST &id query parameter is given,
and it is not accessible, it will fail now.

Resolves: #88357
Releases: master
Change-Id: I3dafec3382d8c6797f32b3acde2763dacd13e0f3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60757
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>