[TASK] Move "Refresh required" modal interception into JavaScript land
The "Refresh required" is currently built in PHP and invoked when changing
a field that is explictly configured as "onChange => reload" in TCA. Since
we aren't limited to the technology of our time anymore, the main logic
can get moved into FormEngine.js and only registration is kept in
FormEngine for now.

This commit also removed some superfluous usages of
`TBE_EDITOR_fieldChanged`, as those are already handled by FormEngine.

Resolves: #88665
Releases: master
Change-Id: I255b70fba7af8253c6bc02611029a74a57a2a043
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61204
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>