[TASK] Deprecate T3_THIS_LOCATION
This patch deprecates the global variable `T3_THIS_LOCATION`. This
variable is mostly used to attach a returnUrl argument to links,
triggered via a redirect on an onclick event.

In most cases this is not necessary anymore, also it's considered bad
practice not to use regular links in an href attribute.

Resolves: #88854
Releases: master
Change-Id: I9d90a59b83b956f0bf1e380a55d6c3a4c2d469b5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61385
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>