[BUGFIX] Set parent uid in newly created IRRE child record
When creating a new IRRE child element the uid of the parent element
was not set. This is a problem when you use an itemsProcFunc for a
select column which relies on the parent's element uid to show
according values. With the parent uid you can query the stored values.

Resolves: #63777
Releases: master, 9.5
Change-Id: I898de29020bf68e25de9ef1d80cc20353a635524
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63294
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Richard Haeser <richard@maxserv.com>