[BUGFIX] Generate valid export url for files in filelist
Since FAL is in place, the export url for files generated in the
clipboard is broken as it takes the full file identifier. Previously it
took the physical file path, the code was never adjusted to be
compatible with FAL.

This patch now resolves the file identifier and generates proper
arguments for export.

Resolves: #88969
Releases: master, 9.5, 8.7
Change-Id: I6599114b19b9d173f88db7beefabe39e90520899
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61749
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>