[BUGFIX] Ensure manually updated slug is saved correctly
This fixes the bug, that after updating the slug and clicking
the toggle URL button, the slug isn't updated correctly.

Therefore, if the value of the input and read only field differ
after clicking the toggle button, now the updated value
of the input field is written into the read only and hidden field.

Furthermore after recreating the slug, the new value is now also
written into the input field to be present after clicking
the toogle URL button again. Otherwise the recreated slug would be
discarded on save.

Resolves: #86664
Releases: master, 9.5
Change-Id: I1cd8652618a7c5f2f7154c499ef1653bfed73c53
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61514
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>