[FEATURE] Make site title configurable per website language
The site title can now be configured with the site module instead of
using the sys_template record. If you don't need localized site titles
you can also set the title on a site base.

The site title in the sys_template record is deprecated now and will be
used as a fallback in v10.

Resolves: #85592
Releases: master
Change-Id: I1caea88ef9bb7c1a4550d68bff1ede91b6f5cb2a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62101
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: Chris Müller <typo3@krue.ml>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Kevin Appelt <kevin.appelt@icloud.com>
Reviewed-by: Chris Müller <typo3@krue.ml>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>