[BUGFIX] Add lang attribute to html tag in Backend
The current language of the TYPO3 Backend user is now added as
attribute 'lang' to the html tag of any module using the ModuleTemplate.

This improves accessibility, as it is used by screen readers that switch
language profiles to provide the correct accent and pronunciation.

In addition, DocumentTemplate behaves the same by adding
proper htmlspecialchars()

Resolves: #89559
Releases: master
Change-Id: I42132e24475b4aa7decbf5173e3717297eda897e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62169
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>