[FEATURE] Allow flexible base url for slug fields in FormEngine
Due to wrong prefixes of TCA type=slug fields for custom records
e.g. in ext:news, a new configuration `prefix` is introduced.

A `userFunc` can be assigned to the new field which then allows to
provide a custom base url based on e.g. the correct detail page.

The default behaviour, which is used for pages, is still applied,
when no custom base url is given.

Resolves: #89573
Releases: master
Change-Id: I529ff26ddeba200b56f0c23aca55daa187c3f40d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62270
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Josef Glatz <josefglatz@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>