[TASK] Use ArrayObject for middleware storage in symfony container
Symfony 4.4 deprecated support for non-object services:
https://github.com/symfony/dependency-injection/commit/6f11b359ab9

The usage of `array` as service is therefore no longer officially
supported in Symfony 5.0. It still works, but in order to make
sure that a possibly introduced type declaration in a 5.x release
doesnt causes a type error, we migrate the middleware service
from array to ArrayObject.

Change-Id: Ie544014b4aa0217b84b0c9d2cc69efde352bef27
Releases: master
Resolves: #89766
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62429
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>