[FEATURE] Introduce HTML emails with Fluid templates
A new FluidEmail class is added that depends on symfony/mime to allow
sending templated emails via Fluid (HTML + Text).

TYPO3 Install Tool ("Test Mail Setup") and the Backend User Login Notification
are migrated to the mail setup, allowing to sending out HTML mails.

It is used for some TYPO3-system-based emails, and can be configured
on a per-installation level.

In addition, a base layout is shipped for TYPO3 messages, so most of the
TYPO3-system emails can now be migrated further-on.

Resolves: #90266
Releases: master
Change-Id: I6961b4ffd7ba39c947f1ce7a85626349c1b4032e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62362
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>