Revert "[BUGFIX] Set tasks backend:lock & backend:unlock as not schedulable"
This reverts commit 26c9a8e32637d506249989318941c923171b7658.

The schedulers *primary* task is to *schedule* tasks, not to execute them
from the backend. The backend-execution is only an additional
functionality, which is available because it's sometime handy for manual
tasks/testing.

We shouldn't mark commands as non-schedulable when scheduling them is a
perectly valid usecase.
Usecases for scheduling backend locks are:
* Planned maintenance locks:
  schedule a backend lock for a planned update or migration where you
  agreed to stop editing as of a certain date.
* Regular edit-locks, e.g.:
  schedule a backend lock for let's say every friday as of 8pm
  schedule backend unlock on mondays at 8am.

The reverted patch tried to fix the situation where the lock task was
executed from the backend scheduler overview. If we'd want to fix that,
the proper fix would be to add another property which states that this
command must not be executed from the backend.

Releases: master, 9.5
Resolves: #89873
Reverts: #89387
Change-Id: I782809ff065c05adba0653aa208d37e1241eb41a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62551
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>