[BUGFIX] Improve visual output of slug proposal
If slugs are generated within Backend forms, the visual output (= label
below slug field) of the generated slug proposal might be broken if no
slash is prepended to the generated slug. This might be the case in custom
database tables where custom slug fields are defined with the
configuration "prependSlash" set to "false" (which is the default value).
It can also be the case if the configuration "prefixParentPageSlug" is set
to "false".

In order to ensure a clean visual output for those labels, the slug
proposalwithin them will now always be prepended by a slash. This makes
sure that the resulting url which is shown within those labels is always
valid. Note that this only effects visual output and is not meant to
modify any slug proposal generated by the SlugHelper class.

Resolves: #90222
Releases: master, 9.5
Change-Id: I93587cfec0db6c3d889fdce60a4515ef47decc82
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63116
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Elias Häußler <elias@haeussler.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Elias Häußler <elias@haeussler.dev>
Reviewed-by: Benni Mack <benni@typo3.org>