[TASK] Restore public and non-internal method signatures
Followup for https://review.typo3.org/c/Packages/TYPO3.CMS/+/63069

The original patch changed both public and non-internal
method signatures and constitutes a breaking change,
but was not committed and documented as a breaking
change. This patch restores the method signatures in
accordance with the non-breaking nature of the original
patch.

Resolves: #90246
Releases: master
Change-Id: Icc832f80aad0466e3bd1555ff32369ef3949b04b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63094
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>