[TASK] Clean up PageLayoutController
With PageLayoutView, the PageLayoutController class
shows various options that are not evaluated anymore,
and thus, can be streamlined to make adjustments easier.

The fields "new_unique_uid", "returnUrl" and "popViewId" is not in use in
the Page Module.

The ContentLayoutPagePositionMap (UX for choosing a new
content element in two steps) is not in use anymore since the
modal box (New Content Element Wizard) was added.

Resolves: #90309
Releases: master
Change-Id: If2c4e8bf3c0bc1a8bc81ce6c9fa417efcef8b428
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63166
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>