[BUGFIX] Prevent empty div in RecordHistory Rollback
In the RecordHistory Rollback template it can happen, that the
historyRow is missing messages about its differences. In this case, an
empty div with a weird looking border would still be visible in the
html. This patch fixes this by wrapping an if condition around the div,
that checks first, if the needed information is available.

Resolves: #90424
Releases: master, 9.5
Change-Id: Icdd5dbf1b1008ab6a64a2b241982450bd385a8f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63314
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>