[BUGFIX] Show fields for files/folders in ElementInformationController
The changes made to the ElementInformationController in #88901 had a few
problems and regressions:

- A lot of information for files and folders, that has previously been
shown, was now missing.
- Metadata for files was missing as well, because this information is
saved in a separate table. (sys_file_metadata)

This patch brings back all the lost information. Also, the Metadata of
a file can now be viewed in the ElementInformation window as well.

Resolves: #90441
Related: #88901
Releases: master
Change-Id: Idb710c83b74845e6ec8b6e2dc7c3d906841ea6ea
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63332
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Eßl <indy.essl@gmail.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Eßl <indy.essl@gmail.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>