[FEATURE] Allow CEs to be saved directly in NewContentElement wizard
When creating content elements via the NewContentElement wizard,
a new option "saveAndClose" can be defined via PageTSconfig
to skip the "FormEngine" part and save a record directly.

This is especially handy for Content Types like "divider" or
for extensions handling with grids.

This saves clicks for the editor, and matched with the
tt_content_defVals it would be possible to create "quick create"
content elements directly via the New Content Element wizard
with pre-defined values.

The feature is enabled by default for the CType "divider".

Resolves: #90461
Releases: master
Change-Id: I5fe171ac1b243c22d4ed9e14f1a5f3c00c088a21
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63341
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: David Steeb <david.steeb@b13.de>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: David Steeb <david.steeb@b13.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>