[BUGFIX] Fallback to default, if BackendLayout can no longer be resolved
In the new fluidbased page module view, if a previously configured
BackendLayout has been removed. (For instance, the extension providing
it has been removed), the module would throw an uncaught error, because
the script would try to call a method on a non-existing BackendLayout.

This patch fixes this problem by automatically falling back to the
"default" BackendLayout, if the requested one couldn't be provided.
This is the same way as the old PageLayoutView would have reacted to
this problem.

Resolves: #90479
Releases: master
Change-Id: I76127e9b53bcb0ac55875857ca11262a323c6013
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63370
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Claus Due <claus@phpmind.net>
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Claus Due <claus@phpmind.net>
Reviewed-by: Kevin Appelt <kevin.appelt@icloud.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>