[TASK] Remove superfluous isset checks
phpstan reported superfluous usages of isset checks for
variables that are defined and whose type is more
specifically checked in the same condition.

Some reported errors of phpstan have been ignored for now
as they are considered false positives and which will be
tackled once there is more type safety in the context of
those reported errors to be able to properly replace the
isset checks.

Releases: master
Resolves: #90576
Change-Id: I028a928bcbc18185f69f7bda4358aaf08cc016b3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63481
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>